% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historical_rolling_volatility.R
\name{hist.vol.interact}
\alias{hist.vol.interact}
\title{Interactive historical daily volatility}
\usage{
hist.vol.interact(ticker, from, to)
}
\arguments{
\item{ticker}{A character vector storing the tickers of the stocks of which you want the
correlation.
Note that since data are downloaded from Yahoo finance, the ticker should be the same as
the ones reported on Yahoo finance. See examples.}

\item{from}{A single character variable storing the starting date from which we want to
download the data. The format should be "yyyy-mm-dd"}

\item{to}{A single character variable storing the ending date up to the day in which we want
to download the stock data. The format should be "yyyy-mm-dd"}
}
\description{
Create an interacrtive graph of the historical daily volatility for the select
stocks based on log-returns.
}
\examples{
\dontrun{
hist.vol.interact(c("AAPL","TSLA","AMZN"), from= "2020-01-01", to= "2022-01-01")
}

}
