% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaron_oscillator.R
\name{interactive.aaron}
\alias{interactive.aaron}
\title{Interactive Aroon oscillator}
\usage{
interactive.aaron(ticker, from, to)
}
\arguments{
\item{ticker}{A single character variable storing the ticker of the stock to be downloaded.
Note that since data are downloaded from Yahoo finance, the ticker should be the same as
the ones reported on Yahoo finance. See examples.}

\item{from}{A single character variable storing the starting date from which we want to
download the data. The format should be "yyyy-mm-dd"}

\item{to}{A single character variable storing the ending date up to the day in which we want
to download the stock data. The format should be "yyyy-mm-dd"}
}
\description{
This function is build to analyze the Aroon oscillator of a stock in a chsosen period
of time. It gives:
\enumerate{
\item The candlestick graph of prices of the stock (open,high,low and close).
\item The interactive graphs of Aroon down, Aroon up, and the Aroon oscillator (difference between Aroon up and down).
\item It creates an HTML table that groups together open prices, closing prices, adjusted prices, high price, low price, daily volume, aaron up, aaron down and volume.
}
}
\examples{
\dontrun{
interactive.aaron("AAPL",from= "2020-01-01", to= "2021-01-01")
}
}
