% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_return.R
\name{plot.return}
\alias{plot.return}
\title{Plot returns of a given stock}
\usage{
plot.return(ticker, from, to, time = "days")
}
\arguments{
\item{ticker}{A single character variable storing the ticker of the stock to be downloaded.
Note that since data are downloaded from Yahoo finance, the ticker should be the same as
the ones reported on Yahoo finance. See examples.}

\item{from}{A single character variable storing the starting date from which we want to
download the data. The format should be "yyyy-mm-dd"}

\item{to}{A single character variable storing the ending date from which we want to
download the data. The format should be "yyyy-mm-dd"}

\item{time}{A single character variable which indicate the desired time frame of the
returns. Should be one among the following: "days" ; "weeks" ; "months" ; "quarters" ;
"years".
Default to "days"}
}
\description{
Plot a time series of the return of a given stock.
}
\examples{
\dontrun{
plot.return( "AAPL"  ,"2016-01-01" ,"2022-01-01" , time = "quarters")

##if you want to store a dataset in the envinroment
AAPL_ret <- plot.return( "AAPL"  ,"2016-01-01" ,"2022-01-01" , time = "quarters")
}
}
