% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFin-package.R
\docType{package}
\name{plotFin-package}
\alias{plotFin}
\alias{plotFin-package}
\title{plotFin: Plot financial graphs}
\description{
This package helps users to have quick financial plots with a simple line code, avoiding to download, clean and plot data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gabrielebonvicini/plotFin}
  \item Report bugs at \url{https://github.com/gabrielebonvicini/plotFin/issues}
}

}
\author{
\strong{Maintainer}: Bianca Pessot \email{bianca.irene.pessot@usi.ch}

Authors:
\itemize{
  \item Gabriele Bonvicini \email{gabriele.bonvicini@usi.ch}
  \item Vehbian Kraja \email{vehbian.kraja@usi.ch}
  \item Matilde Bergna \email{matilde.bergna@usi.ch}
}

}
\keyword{internal}
