% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stocksharpe_ratio.R
\name{stock.sharpe.ratio}
\alias{stock.sharpe.ratio}
\title{Sharpe Ratio plot}
\usage{
stock.sharpe.ratio(ticker, from, to, risk_free, weights)
}
\arguments{
\item{ticker}{A character vector of length two storing  tickers of the stocks of
which you want the z-score
Note that since data are downloaded from Yahoo finance, the ticker should be the same as
the ones reported on Yahoo finance. See examples.}

\item{from}{A single character variable storing the starting date from which we want to
download the data. The format should be "yyyy-mm-dd"}

\item{to}{A single character variable storing the ending date from which we want to
download the data. The format should be "yyyy-mm-dd"}

\item{risk_free}{Numerical variable of length one. It indicates the risk free rate}

\item{weights}{Numerical vector of length two. It indicates the weight that each stock has}
}
\description{
It is a function that represents the sharpe ratio of a portfolio composed of two stocks
in a given data frame. The Sharpe index is a metric that measures the extra return
realized by a portfolio (or fund) per unit of total risk borne in comparison to the
risk-free rate. This function has been created calculating the ratio between the
difference between the  portfolio return and risk free rate and  the standard
deviation of the portfolio.
}
\examples{
\dontrun{
stock.sharpe.ratio( ticker = c("NFLX","AMZN"), weights=c(0.50,0.50),
from="2010-01-01", to="2022-01-01", risk_free = 0.05)
}
}
