% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_score.R
\name{z.value}
\alias{z.value}
\title{Z-score comparison}
\usage{
z.value(ticker, from, to)
}
\arguments{
\item{ticker}{A character vector of length two storing  tickers of the stocks of
which you want the z-score
Note that since data are downloaded from Yahoo finance, the ticker should be the same as
the ones reported on Yahoo finance. See examples.}

\item{from}{A single character variable storing the starting date from which we want to
download the data. The format should be "yyyy-mm-dd"}

\item{to}{A single character variable storing the ending date from which we want to
download the data. The format should be "yyyy-mm-dd"}
}
\description{
A function  that makes comparison between the z-value of two stocks,
in a monthly time frame. The value of the z-score shows how many standard deviations
there are away from the mean. It has been calculated using the difference between
the return of each ticket and its mean, divided by its standard deviation.
}
\examples{
\dontrun{ 
z.value(c("AAPL",  "MSFT"), from = "2020-01-01", to = "2021-06-06")
}

}
